pyproject.toml
[project]
name = "dopaminas-loja"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "aiofiles",
    "aiohttp",
    "cachetools",
    "httpx",
    "pydantic",
    "pymongo",
    "pytelegrambotapi",
    "python-decouple",
    "stripe",
]
